Problem: Postgres enum value validity rules
They are wider than those for GraphQL as mentioned in #172.

Solution: allow specifying name mapping for enums via comments

Creating a comment like:

```postgresql
comment on type my_enum is E'@graphql({"mappings": {"valid value": "valid_value"}})';
```

Allows to override Postgres enum value to have a different value.

Please note that this is the first go at the problem, and it doesn't
handle the following (and probably more):

* [ ] verification of new value's validity
* [ ] mutations
* [ ] more exhaustive testing

Also, it is important to mention that while this approach does not
implement a dynamic mappin technique proposed in #172 for the sake of
conformity to the current approach used throughout pg_graphsql, it does
not mean this will not or should happen in the future. A breaking
release may indeed consider such a change for a variety of reasons:

* performance optimizations (mapping functions can be declared immutable)
* cases where comments aren't feasible or too onerous

Of particular note is an idea to dispatch conversions back to Postgres
using an `inout` parameter to help us to type on the return type.