Problem: duplication in SQL context
We insert enum mappings  into SQL context twice to be able to get them
at the point where we need it. This is suboptimal.

Solution: mark this duplication as temporary with TODO

As per
https://github.com/supabase/pg_graphql/pull/376#discussion_r1259865044,
this is to be superseded by a better internal design:

For example, what if instead of storing type_oid as u32 across the code
base and if we refactor Type/TypeCategory to contain a reference to the
actual type (Arc<Enum>, Arc<Composite>, etc.), then at any point, we
can look up the type.