Problem: preparing type information for some Array types fails
Particularly, we can no longer get a mutable reference to Arc<Type>
of oidvector, int2vector.

Solution: ensure we hydrate typing information for referred eypes first

This way, we can't run into a situation where we have a type that refers
to, say, oidvector but we haven't hydrated oidvector yet. Previously, by
the time we reached oidvector, it was already referred to elsewhere and
thus the Arc was no longer available mutably.