Problem: brittle element removal logic in the type hydration pass
If for some reason we mess this up in the future and don't return the
type to list, it may be hard to detect this error.

Solution: be defensive and check that we do put it back.

Background: the reason we remove it from the list is just to get around
borrow-checker rules as we can't borrow context.types twice.