fix failing char(x) test
The test failed because a function of type char(x) was being cast to
`character` instead of `char(x)`. This was because when we load the sql
context we do not have the type modifier info to find out what `x` in
char(`x`) is for function inputs. As a workaround we cast such vales to
text instead.