cast oids to bigint instead of int to avoid overflow
oids are 4 byte unsigned integers. ints are 4 bytes signed integers. This means
some oids when cast to int are represented as negative values which are saved in
the context json. When the Rust code tries to parse these negative values as a
u32 (which is also an unsigned 4 byte value) it fails. Casting to bigint works
around this because bigint is an 8 byte long signed integer and all of the oids
fit withing the positive values range of a bigint.